+++ front-end/src/layout/MainLayout/Header/WelcomeSection/index.js
@@ -1,10 +1,16 @@
-import { useSelector } from 'react-redux';
+import { shallowEqual, useSelector } from 'react-redux';
+import { useState, useEffect } from 'react';
 
 const WelcomeMessage = () => {
-    let user = useSelector((state) => state.user);
-    console.log(user);
+    let [firstName, setFirstName] = useState('Bison');
+    let user = useSelector((state) => state.user, shallowEqual);
+    useEffect(() => {
+        if (user.firstName !== '') {
+            setFirstName(user.firstName);
+        }
+    }, user);
 
-    return <h1 style={{ paddingLeft: '25px' }}> Welcome, {user.firstName}!</h1>;
+    return <h1 style={{ paddingLeft: '25px' }}> Welcome, {firstName}!</h1>;
 };
 
 export default WelcomeMessage;
