--- front-end/src/layout/MinimalLayout/index.js
@@ -1,11 +1,0 @@
-import { Outlet } from 'react-router-dom';
-
-// ==============================|| MINIMAL LAYOUT ||============================== //
-
-const MinimalLayout = () => (
-    <>
-        <Outlet />
-    </>
-);
-
-export default MinimalLayout;
