+++ front-end/src/routes/AuthenticationRoutes.js
@@ -2,7 +2,6 @@ import { lazy } from 'react';
 
 // project imports
 import Loadable from 'ui-component/Loadable';
-import MinimalLayout from 'layout/MinimalLayout';
 
 // login option 3 routing
 const AuthLogin = Loadable(lazy(() => import('views/pages/authentication/auth-template/Login')));
@@ -12,7 +11,6 @@ const AuthTransferMoney = Loadable(lazy(() => import('views/pages/authentication
 
 const AuthenticationRoutes = {
     path: '/',
-    element: <MinimalLayout />,
     children: [
         {
             path: '/user/login',
