--- front-end/src/ui-component/cards/Skeleton/ImagePlaceholder.js
@@ -1,8 +1,0 @@
-// material-ui
-import Skeleton from '@mui/material/Skeleton';
-
-// ==============================|| SKELETON IMAGE CARD ||============================== //
-
-const ImagePlaceholder = ({ ...others }) => <Skeleton variant="rectangular" {...others} animation="wave" />;
-
-export default ImagePlaceholder;
