+++ front-end/src/views/dashboard/Default/EarningCard.js
@@ -1,5 +1,5 @@
-import PropTypes from 'prop-types';
 import { useState } from 'react';
+import { shallowEqual, useSelector } from 'react-redux';
 
 // material-ui
 import { styled, useTheme } from '@mui/material/styles';
@@ -7,12 +7,12 @@ import { Avatar, Box, Grid, Menu, MenuItem, Typography } from '@mui/material';
 
 // project imports
 import MainCard from 'ui-component/cards/MainCard';
-import SkeletonEarningCard from 'ui-component/cards/Skeleton/EarningCard';
+// import SkeletonEarningCard from 'ui-component/cards/Skeleton/EarningCard';
 
 // assets
-import EarningIcon from 'assets/images/icons/earning.svg';
+import EarningIcon from 'assets/images/earning.svg';
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz';
-import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
+import AccountBalanceWalletIcon from '@mui/icons-material/AccountBalanceWallet';
 import GetAppTwoToneIcon from '@mui/icons-material/GetAppOutlined';
 import FileCopyTwoToneIcon from '@mui/icons-material/FileCopyOutlined';
 import PictureAsPdfTwoToneIcon from '@mui/icons-material/PictureAsPdfOutlined';
@@ -56,7 +56,8 @@ const CardWrapper = styled(MainCard)(({ theme }) => ({
 
 // ===========================|| DASHBOARD DEFAULT - EARNING CARD ||=========================== //
 
-const EarningCard = ({ isLoading }) => {
+const EarningCard = () => {
+    let accountBalance = useSelector((state) => state.user.accountBalance, shallowEqual);
     const theme = useTheme();
 
     const [anchorEl, setAnchorEl] = useState(null);
@@ -71,117 +72,109 @@ const EarningCard = ({ isLoading }) => {
 
     return (
         <>
-            {isLoading ? (
-                <SkeletonEarningCard />
-            ) : (
-                <CardWrapper border={false} content={false}>
-                    <Box sx={{ p: 2.25 }}>
-                        <Grid container direction="column">
-                            <Grid item>
-                                <Grid container justifyContent="space-between">
-                                    <Grid item>
-                                        <Avatar
-                                            variant="rounded"
-                                            sx={{
-                                                ...theme.typography.commonAvatar,
-                                                ...theme.typography.largeAvatar,
-                                                backgroundColor: theme.palette.secondary[800],
-                                                mt: 1
-                                            }}
-                                        >
-                                            <img src={EarningIcon} alt="Notification" />
-                                        </Avatar>
-                                    </Grid>
-                                    <Grid item>
-                                        <Avatar
-                                            variant="rounded"
-                                            sx={{
-                                                ...theme.typography.commonAvatar,
-                                                ...theme.typography.mediumAvatar,
-                                                backgroundColor: theme.palette.secondary.dark,
-                                                color: theme.palette.secondary[200],
-                                                zIndex: 1
-                                            }}
-                                            aria-controls="menu-earning-card"
-                                            aria-haspopup="true"
-                                            onClick={handleClick}
-                                        >
-                                            <MoreHorizIcon fontSize="inherit" />
-                                        </Avatar>
-                                        <Menu
-                                            id="menu-earning-card"
-                                            anchorEl={anchorEl}
-                                            keepMounted
-                                            open={Boolean(anchorEl)}
-                                            onClose={handleClose}
-                                            variant="selectedMenu"
-                                            anchorOrigin={{
-                                                vertical: 'bottom',
-                                                horizontal: 'right'
-                                            }}
-                                            transformOrigin={{
-                                                vertical: 'top',
-                                                horizontal: 'right'
-                                            }}
-                                        >
-                                            <MenuItem onClick={handleClose}>
-                                                <GetAppTwoToneIcon sx={{ mr: 1.75 }} /> Import Card
-                                            </MenuItem>
-                                            <MenuItem onClick={handleClose}>
-                                                <FileCopyTwoToneIcon sx={{ mr: 1.75 }} /> Copy Data
-                                            </MenuItem>
-                                            <MenuItem onClick={handleClose}>
-                                                <PictureAsPdfTwoToneIcon sx={{ mr: 1.75 }} /> Export
-                                            </MenuItem>
-                                            <MenuItem onClick={handleClose}>
-                                                <ArchiveTwoToneIcon sx={{ mr: 1.75 }} /> Archive File
-                                            </MenuItem>
-                                        </Menu>
-                                    </Grid>
+            <CardWrapper border={false} content={false}>
+                <Box sx={{ p: 2.25 }}>
+                    <Grid container direction="column">
+                        <Grid item>
+                            <Grid container justifyContent="space-between">
+                                <Grid item>
+                                    <Avatar
+                                        variant="rounded"
+                                        sx={{
+                                            ...theme.typography.commonAvatar,
+                                            ...theme.typography.largeAvatar,
+                                            backgroundColor: theme.palette.secondary[800],
+                                            mt: 1
+                                        }}
+                                    >
+                                        <img src={EarningIcon} alt="Notification" />
+                                    </Avatar>
                                 </Grid>
-                            </Grid>
-                            <Grid item>
-                                <Grid container alignItems="center">
-                                    <Grid item>
-                                        <Typography sx={{ fontSize: '2.125rem', fontWeight: 500, mr: 1, mt: 1.75, mb: 0.75 }}>
-                                            $500.00
-                                        </Typography>
-                                    </Grid>
-                                    <Grid item>
-                                        <Avatar
-                                            sx={{
-                                                cursor: 'pointer',
-                                                ...theme.typography.smallAvatar,
-                                                backgroundColor: theme.palette.secondary[200],
-                                                color: theme.palette.secondary.dark
-                                            }}
-                                        >
-                                            <ArrowUpwardIcon fontSize="inherit" sx={{ transform: 'rotate3d(1, 1, 1, 45deg)' }} />
-                                        </Avatar>
-                                    </Grid>
+                                <Grid item>
+                                    <Avatar
+                                        variant="rounded"
+                                        sx={{
+                                            ...theme.typography.commonAvatar,
+                                            ...theme.typography.mediumAvatar,
+                                            backgroundColor: theme.palette.secondary.dark,
+                                            color: theme.palette.secondary[200],
+                                            zIndex: 1
+                                        }}
+                                        aria-controls="menu-earning-card"
+                                        aria-haspopup="true"
+                                        onClick={handleClick}
+                                    >
+                                        <MoreHorizIcon fontSize="inherit" />
+                                    </Avatar>
+                                    <Menu
+                                        id="menu-earning-card"
+                                        anchorEl={anchorEl}
+                                        keepMounted
+                                        open={Boolean(anchorEl)}
+                                        onClose={handleClose}
+                                        variant="selectedMenu"
+                                        anchorOrigin={{
+                                            vertical: 'bottom',
+                                            horizontal: 'right'
+                                        }}
+                                        transformOrigin={{
+                                            vertical: 'top',
+                                            horizontal: 'right'
+                                        }}
+                                    >
+                                        <MenuItem onClick={handleClose}>
+                                            <GetAppTwoToneIcon sx={{ mr: 1.75 }} /> Import Card
+                                        </MenuItem>
+                                        <MenuItem onClick={handleClose}>
+                                            <FileCopyTwoToneIcon sx={{ mr: 1.75 }} /> Copy Data
+                                        </MenuItem>
+                                        <MenuItem onClick={handleClose}>
+                                            <PictureAsPdfTwoToneIcon sx={{ mr: 1.75 }} /> Export
+                                        </MenuItem>
+                                        <MenuItem onClick={handleClose}>
+                                            <ArchiveTwoToneIcon sx={{ mr: 1.75 }} /> Archive File
+                                        </MenuItem>
+                                    </Menu>
                                 </Grid>
-                            </Grid>
-                            <Grid item sx={{ mb: 1.25 }}>
-                                <Typography
-                                    sx={{
-                                        fontSize: '1rem',
-                                        fontWeight: 500,
-                                        color: theme.palette.secondary[200]
-                                    }}
-                                >
-                                    Total Earning
-                                </Typography>
                             </Grid>
                         </Grid>
-                    </Box>
-                </CardWrapper>
-            )}
+                        <Grid item>
+                            <Grid container alignItems="center">
+                                <Grid item>
+                                    <Typography sx={{ fontSize: '2.125rem', fontWeight: 500, mr: 1, mt: 1.75, mb: 0.75 }}>
+                                        ${accountBalance}
+                                    </Typography>
+                                </Grid>
+                                <Grid item>
+                                    <Avatar
+                                        sx={{
+                                            cursor: 'pointer',
+                                            ...theme.typography.smallAvatar,
+                                            backgroundColor: theme.palette.secondary[200],
+                                            color: theme.palette.secondary.dark
+                                        }}
+                                    >
+                                        <AccountBalanceWalletIcon fontSize="inherit" />
+                                    </Avatar>
+                                </Grid>
+                            </Grid>
+                        </Grid>
+                        <Grid item sx={{ mb: 1.25 }}>
+                            <Typography
+                                sx={{
+                                    fontSize: '1rem',
+                                    fontWeight: 500,
+                                    color: theme.palette.secondary[200]
+                                }}
+                            >
+                                Total Earning
+                            </Typography>
+                        </Grid>
+                    </Grid>
+                </Box>
+            </CardWrapper>
         </>
     );
-};
-
-EarningCard.propTypes = {
-    isLoading: PropTypes.bool
 };
 
 export default EarningCard;
