--- front-end/src/views/dashboard/Default/ExtraCard.js
@@ -1,116 +1,0 @@
-import PropTypes from 'prop-types';
-import { useState } from 'react';
-
-// material-ui
-import { useTheme, styled } from '@mui/material/styles';
-import { Avatar, Box, Button, Grid, Typography } from '@mui/material';
-
-// third-party
-import Chart from 'react-apexcharts';
-
-// project imports
-import MainCard from 'ui-component/cards/MainCard';
-import SkeletonTotalOrderCard from 'ui-component/cards/Skeleton/EarningCard';
-
-import ChartDataMonth from './chart-data/total-order-month-line-chart';
-import ChartDataYear from './chart-data/total-order-year-line-chart';
-
-// assets
-import LocalMallOutlinedIcon from '@mui/icons-material/LocalMallOutlined';
-import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
-import chartData from './chart-data/bajaj-area-chart';
-
-const CardWrapper = styled(MainCard)(({ theme }) => ({
-    backgroundColor: 'red',
-    color: '#fff',
-    overflow: 'hidden',
-    position: 'relative',
-    height: 185,
-    '&>div': {
-        position: 'relative',
-        zIndex: 5
-    },
-    '&:after': {
-        content: '""',
-        position: 'absolute',
-        width: 210,
-        height: 210,
-        background: theme.palette.primary[800],
-        borderRadius: '50%',
-        zIndex: 1,
-        top: -85,
-        right: -95,
-        [theme.breakpoints.down('sm')]: {
-            top: -105,
-            right: -140
-        }
-    },
-    '&:before': {
-        content: '""',
-        position: 'absolute',
-        zIndex: 1,
-        width: 210,
-        height: 210,
-        background: theme.palette.primary[800],
-        borderRadius: '50%',
-        top: -125,
-        right: -15,
-        opacity: 0.5,
-        [theme.breakpoints.down('sm')]: {
-            top: -155,
-            right: -70
-        }
-    }
-}));
-
-// ==============================|| DASHBOARD - TOTAL ORDER LINE CHART CARD ||============================== //
-
-const ExtraCard = ({ isLoading }) => {
-    const theme = useTheme();
-
-    const [timeValue, setTimeValue] = useState(false);
-    const handleChangeTime = (event, newValue) => {
-        setTimeValue(newValue);
-    };
-
-    return (
-        <>
-            {isLoading ? (
-                <SkeletonTotalOrderCard />
-            ) : (
-                <CardWrapper border={false} content={false}>
-                    <Box sx={{ p: 2.25 }}>
-                        <Grid container direction="column">
-                            <Grid item sx={{ mb: 0.75 }}>
-                                <Grid container alignItems="center">
-                                    <Grid item xs={6}>
-                                        <Grid item xs={12}>
-                                            <Typography
-                                                sx={{
-                                                    fontSize: '1rem',
-                                                    fontWeight: 500,
-                                                    color: theme.palette.primary[200]
-                                                }}
-                                            >
-                                                Transaction History
-                                            </Typography>
-                                        </Grid>
-                                    </Grid>
-                                </Grid>
-                            </Grid>
-                        </Grid>
-                        <Grid item xs={6}>
-                            <Chart {...chartData} />
-                        </Grid>
-                    </Box>
-                </CardWrapper>
-            )}
-        </>
-    );
-};
-
-ExtraCard.propTypes = {
-    isLoading: PropTypes.bool
-};
-
-export default ExtraCard;
