+++ front-end/src/views/dashboard/Default/index.js
@@ -1,42 +1,43 @@
-import { useEffect, useState } from 'react';
-
 // material-ui
 import { Grid } from '@mui/material';
 
 // project imports
 import EarningCard from './EarningCard';
 import PopularCard from './PopularCard';
-import TotalOrderLineChartCard from './TotalOrderLineChartCard';
+import TransactionCard from './TransactionCard';
+import DarkCard from './DarkCard';
+import LightCard from './LightCard';
 import { gridSpacing } from 'store/constant';
-import ExtraCard from './ExtraCard';
 
 // ==============================|| DEFAULT DASHBOARD ||============================== //
 
 const Dashboard = () => {
-    const [isLoading, setLoading] = useState(true);
-    useEffect(() => {
-        setLoading(false);
-    }, []);
-
     return (
         <Grid container spacing={gridSpacing}>
             <Grid item xs={12}>
                 <Grid container spacing={gridSpacing}>
                     <Grid item lg={4} md={6} sm={6} xs={12}>
-                        <EarningCard isLoading={isLoading} />
+                        <EarningCard />
                     </Grid>
                     <Grid item lg={4} md={6} sm={6} xs={12}>
-                        <TotalOrderLineChartCard isLoading={isLoading} />
+                        <TransactionCard />
                     </Grid>
-                    <Grid item lg={4} md={6} sm={6} xs={12}>
-                        <ExtraCard isLoading={isLoading} />
+                    <Grid item lg={4} md={12} sm={12} xs={12}>
+                        <Grid container spacing={gridSpacing}>
+                            <Grid item sm={6} xs={12} md={6} lg={12}>
+                                <DarkCard />
+                            </Grid>
+                            <Grid item sm={6} xs={12} md={6} lg={12}>
+                                <LightCard />
+                            </Grid>
+                        </Grid>
                     </Grid>
                 </Grid>
             </Grid>
             <Grid item xs={12}>
                 <Grid container spacing={gridSpacing}>
                     <Grid item xs={12} md={12}>
-                        <PopularCard isLoading={isLoading} />
+                        <PopularCard />
                     </Grid>
                 </Grid>
             </Grid>
