+++ front-end/src/views/pages/authentication/userSlice.js
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-let initialState = {
+let blankState = {
     userName: '',
     firstName: '',
     lastName: '',
@@ -9,14 +9,24 @@ let initialState = {
     transferHistory: []
 };
 
-// ==============================|| CUSTOMIZATION REDUCER ||============================== //
+const userInfo = localStorage.getItem('userInfo') !== null ? JSON.parse(localStorage.getItem('userInfo')) : blankState;
+
+let initialState = {
+    userName: userInfo.userName,
+    firstName: userInfo.firstName,
+    lastName: userInfo.lastName,
+    accountBalance: userInfo.accountBalance,
+    expenseHistory: userInfo.expenseHistory,
+    transferHistory: userInfo.transferHistory
+};
 
 export let userSlice = createSlice({
     name: 'user',
     initialState,
     reducers: {
         setUser: (state, action) => {
-            state = action.payload.user;
+            state = action.payload.userInfo;
+            localStorage.setItem('userInfo', JSON.stringify(state));
             return state;
         }
     }
